Gasoline price vs Ethanol
This code compares the price of gasoline with the price of ethanol. It calculates the adjusted price of gasoline (gasoline * 0.7) and makes the following recommendations:
If the adjusted gasoline price is lower than the ethanol price, it recommends using gasoline.
If it is higher, it recommends using ethanol.
If the prices are the same, it indicates that it doesn’t matter which one to use.